fix: separate currency input
